Adding the real id of the user we want to follow
Before, when saving the list of followed user with the newly followed user, it was creating a unique object id for this since it didn't had one. By setting the _id on the newly added user to the list of following, it will not generate an object id and will use the _id. 

It will let use the id for visiting friends of our friends and .... inception!